SE2 tests

+++ lib/src/dgsvj0.dart
@@ -762,9 +762,7 @@ void dgsvj0(
 
     if ((i < SWBAND) && ((MXAAPQ <= ROOTTOL) || (ISWROT <= N))) SWBAND = i;
 
-    if ((i > SWBAND + 1) &&
-        (MXAAPQ < N.toDouble() * TOL) &&
-        (N.toDouble() * MXAAPQ * MXSINJ < TOL)) {
+    if ((i > SWBAND + 1) && (MXAAPQ < N * TOL) && (N * MXAAPQ * MXSINJ < TOL)) {
       isBelowTolerance = true;
       break;
     }
