+++ lib/src/dlaruv.dart
@@ -173,9 +173,7 @@ void dlaruv(
 
       // Convert 48-bit integer to a real number in the interval (0,1)
 
-      X[I] = R *
-          (IT1.toDouble() +
-              R * (IT2.toDouble() + R * (IT3.toDouble() + R * IT4.toDouble())));
+      X[I] = R * (IT1 + R * (IT2 + R * (IT3 + R * IT4)));
 
       if (X[I] != 1.0) break;
