+++ lib/src/zlaqr3.dart
@@ -129,7 +129,7 @@ void zlaqr3(
   SAFMIN = dlamch('SAFE MINIMUM');
   // SAFMAX = RONE / SAFMIN;
   ULP = dlamch('PRECISION');
-  SMLNUM = SAFMIN * (N.toDouble() / ULP);
+  SMLNUM = SAFMIN * (N / ULP);
 
   // ==== Setup deflation window ====
